+++ src/components/hooks/useTodos.tsx
@@ -1,5 +1,5 @@
 import React, { FormEvent, useState } from 'react';
-import { AllToDos, defaultToDo, ToDo } from 'components/@utils/type';
+import { AllToDos, defaultToDo, ToDo } from 'components/utils/type';
 
 export const useTodos = () => {
   const [allToDos, setAllToDos] = useState<AllToDos>([]);
@@ -60,31 +60,5 @@ export const useTodos = () => {
     setAllToDos,
     restToDos,
     todo,
-  };
-};
-
-export const useModal = (allToDos: AllToDos) => {
-  const [modalShow, setModalShow] = useState(false);
-  const [currentTodo, setCurrentToDo] = useState('');
-  const [currentTodoObj, setCurrentTodoObj] = useState<ToDo>({ ...defaultToDo });
-
-  const openModal = () => setModalShow(true);
-  const closeModal = () => setModalShow(false);
-  const modalHandler = (e: React.MouseEvent) => {
-    const { id: selectedId } = e.currentTarget;
-    const selectedToDo = allToDos.find((toDo) => toDo.id === selectedId);
-
-    openModal();
-    setCurrentToDo(selectedToDo?.todo ?? '');
-    setCurrentTodoObj(selectedToDo ?? { ...defaultToDo });
-  };
-
-  return {
-    modalHandler,
-    setCurrentToDo,
-    currentTodoObj,
-    currentTodo,
-    modalShow,
-    closeModal,
   };
 };
